Implemented: Init updater with opts and it works independently. Calls onUpdateReady when update has been downloaded and extracted. Checks for update once a day or every opts.updateCheckInterval ms

+++ src/index.js
@@ -8,7 +8,16 @@ const fork = require('child_process').fork;
 const superagent = require('superagent');
 const { promisify } = require('util');
 const untar = promisify(require('targz').decompress);
-const unlink = promisify(require('fs').unlink);
+const rimraf = promisify(require('rimraf'));
+const fs = require('fs');
+const compareVersions = require('compare-versions');
+const isSemver = require('is-semver');
+
+const unlink = promisify(fs.unlink);
+const readdir = promisify(fs.readdir);
+const stat = promisify(fs.stat);
+
+const defaultUpdateCheckInterval = (1000 * 60 * 60 * 24) * 1; //Check every 1 days
 
 function downloadUpdateFile(url, dir, checksum) {
     return new Promise((resolve, reject) => {
@@ -26,6 +35,36 @@ function downloadUpdateFile(url, dir, checksum) {
                 return reject(message);
         });
     });
+};
+
+function setupUpdateChecker(opts) {
+    const updateCheckInterval = opts.updateCheckInterval || defaultUpdateCheckInterval;
+    setInterval(updateIfAvailable.bind(null, opts), updateCheckInterval); //Setup every x ms check
+    setImmediate(updateIfAvailable.bind(null, opts)); //Setup immediate check
+};
+
+async function getAvailableSemvers(opts) {
+    const itemsInDir = await readdir(opts.dirPath);
+    const areItemsDirs = itemsInDir.map(item => stat(path.join(opts.dirPath, `./${item}`)));
+
+    const stats = await Promise.all(areItemsDirs);
+    let dirs = [];
+    let files = [];
+
+    itemsInDir.forEach((item, i) => {
+        if (stats[i].isDirectory() && isSemver(item))
+            dirs.push(item);
+        else if (stats[i].isFile())
+            files.push(item);
+    });
+
+    const availableSemvers = dirs.filter(dirName => files.indexOf(`${dirName}.tar.gz`) == -1);
+    return availableSemvers.sort(compareVersions);
+};
+
+async function getLatestAvailableSemver(opts) {
+    const availableSemvers = await getAvailableSemvers(opts);
+    return availableSemvers[availableSemvers.length - 1];
 };
 
 async function updateIfAvailable(opts) {
@@ -50,17 +89,35 @@ async function updateIfAvailable(opts) {
         }
     }
 
+    const semver = opts.getSemver(updateJSON);
+    const availableSemvers = await getAvailableSemvers(opts);
+    const isVersionAlreadyDownloaded = availableSemvers.indexOf(semver) > -1;
+
     const shouldDownload = opts.shouldDownload(updateJSON);
+
+    if (isVersionAlreadyDownloaded) {
+        return debug(`No updates available. Quitting update check.`);
+    }
+
     if (shouldDownload instanceof Promise ? await shouldDownload : shouldDownload) {
         let updateFilePath;
         const downloadURL = opts.getDownloadURL(updateJSON);
 
         try {
-            updateFilePath = await downloadUpdateFile(downloadURL, opts.dir, opts.getChecksum(updateJSON));
+            updateFilePath = await downloadUpdateFile(downloadURL, opts.dirPath, opts.getChecksum(updateJSON));
+
+            const extractDir = path.join(opts.dirPath, `./${semver}`);
+            debug(`Extracting update targz to ${extractDir}`);
+
+            await rimraf(extractDir);
             await untar({
                 src: updateFilePath,
-                dest: opts.extractDir
+                dest: extractDir
             });
+            await unlink(updateFilePath);
+
+            if (opts.onUpdateReady)
+                return opts.onUpdateReady(semver);
         } catch (err) {
             if (updateFilePath)
                 await unlink(updateFilePath);
@@ -71,18 +128,38 @@ async function updateIfAvailable(opts) {
     }
 };
 
-async function test() {
-    const downloadFilePath = await updateIfAvailable({
-        shouldDownload: () => true,
-        dir: path.join(__dirname, './tmp'),
-        url: 'http://localhost:3002/update.json',
-        getDownloadURL: manifest => manifest.url,
-        extractDir: path.join(__dirname, './tmp'),
-        getChecksum: manifest => manifest.checksum
-    });
+async function getExecutable(opts) {
+    const semver = opts.version || await getLatestAvailableSemver(opts);
+    const dirPath = path.join(opts.dirPath, `./${semver}`);
+
+    debug(`Requiring version ${semver} from ${dirPath}`);
+    return {
+        requirePath: dirPath,
+        semver: semver
+    };
 };
 
-//test();
+async function test() {
+    const { requirePath, semver } = await init({
+        shouldDownload: () => true,
+        dirPath: path.join(__dirname, './tmp'),
+        url: 'http://localhost:3002/update.json',
+        getDownloadURL: manifest => manifest.url,
+        getChecksum: manifest => manifest.checksum,
+        getSemver: manifest => manifest.semver,
+        onUpdateReady: semver => debug(`Update version ${semver} Extracted and Ready to go.`)
+    });
+
+    debug(`requirePath: ${requirePath}`);
+    debug(`semver: ${semver}`);
+};
+
+test();
+
+async function init(opts) {
+    setupUpdateChecker(opts);
+    return await getExecutable(opts);
+};
 
 module.exports = {
     updateIfAvailable
